Add confirm dialog before deleting (#44)

+++ src/views/BackOffice/Categories.js
@@ -30,18 +30,20 @@ export default function Categories() {
     }
 
     async function deleteCategory(id) {
-        setLockedCategoryIds(state => [...state, id]);
+        if(window.confirm('¿Realmente desea eliminar esta categoría?')) {
+            setLockedCategoryIds(state => [...state, id]);
 
-        const { success, errorMessage } = await deleteCategoryService(id);
+            const { success, errorMessage } = await deleteCategoryService(id);
 
-        if (success) {
-            // Remove deleted category from categories array.
-            setCategories(state => state.filter(category => category.id !== id));
-        } else {
-            toast.error('Error deleting category: ' + errorMessage);
+            if (success) {
+                // Remove deleted category from categories array.
+                setCategories(state => state.filter(category => category.id !== id));
+            } else {
+                toast.error('Error deleting category: ' + errorMessage);
+            }
+
+            setLockedCategoryIds(state => state.filter(categoryId => categoryId !== id));
         }
-
-        setLockedCategoryIds(state => state.filter(categoryId => categoryId !== id));
     }
 
     return (
